pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "optionpilot"
version = "0.1.0"
description = "AI Options Strategy Analyzer - MVP desktop application for options trading analysis"
authors = [{name = "OptionPilot Team"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.25.0",
    "pandas>=2.0.0", 
    "python-dotenv>=1.0.0",
    "matplotlib>=3.7.0",
    "streamlit>=1.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

[project.scripts]
optionpilot = "cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"